+++ raphael-solver/src/step_lower_bound_solver/solver.rs
@@ -39,6 +39,7 @@ pub struct StepLbSolver {
 
 impl StepLbSolver {
     pub fn new(mut settings: SolverSettings, interrupt_signal: utils::AtomicFlag) -> Self {
+        settings.simulator_settings.adversarial = false;
         ReducedState::optimize_action_mask(&mut settings.simulator_settings);
         Self {
             settings,
@@ -51,7 +52,7 @@ impl StepLbSolver {
             precompute_templates: Self::generate_precompute_templates(&settings),
             next_precompute_step_budget: NonZeroU8::new(1).unwrap(),
             precomputed_states: 0,
-            iq_quality_lut: compute_iq_quality_lut(&settings),
+            iq_quality_lut: utils::compute_iq_quality_lut(&settings),
         }
     }
 
@@ -64,7 +65,7 @@ impl StepLbSolver {
             effects: Effects::initial(&settings.simulator_settings)
                 .with_quick_innovation_available(false)
                 .with_heart_and_soul_available(false)
-                .with_adversarial_guard(true)
+                .with_adversarial_guard(false)
                 .with_combo(Combo::None),
         };
         templates.insert(seed_template);
@@ -360,35 +361,4 @@ impl Template {
         };
         ReducedState::from_state(state, step_budget)
     }
-}
-
-fn compute_iq_quality_lut(settings: &SolverSettings) -> [u32; 11] {
-    let mut result = [u32::MAX; 11];
-    result[0] = 0;
-    for iq in 0..10 {
-        let state = SimulationState {
-            cp: 500,
-            durability: 100,
-            progress: 0,
-            quality: 0,
-            unreliable_quality: 0,
-            effects: Effects::new()
-                .with_allow_quality_actions(true)
-                .with_adversarial_guard(true)
-                .with_inner_quiet(iq),
-        };
-        for &action in FULL_SEARCH_ACTIONS {
-            if let Ok(new_state) = use_action_combo(settings, state, action) {
-                let new_iq = new_state.effects.inner_quiet();
-                if new_iq > iq {
-                    let action_quality = new_state.quality;
-                    result[usize::from(new_iq)] = std::cmp::min(
-                        result[usize::from(new_iq)],
-                        result[usize::from(iq)] + action_quality,
-                    );
-                }
-            }
-        }
-    }
-    result
 }
